[fix] #6 HoldingDto의 평균 매입가 메서드를 double로 수정

+++ src/main/java/farmsystem/backend/domain/portfolio/dto/HoldingDto.java
@@ -7,7 +7,7 @@ public record HoldingDto(
         long buyPrice,
         int buyAmount
 ) {
-    public long getAveragePrice() {
-        return buyAmount == 0 ? 0 : buyPrice / buyAmount;
+    public double getAveragePrice() {
+        return buyAmount == 0 ? 0 : (double) buyPrice / buyAmount;
     }
 }
